// ui/packages/src/api/generated/filesys/schemas/fileShareAccessRecordDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Refactored free-fs with a subject-match-java style layered architecture.
 * OpenAPI spec version: v1.0.0
 */

export interface FileShareAccessRecordDTO {
  /** 访问地理位置 */
  accessAddress?: string;
  /** 访问IP地址 */
  accessIp?: string;
  /** 访问时间 */
  accessTime?: string;
  /** 浏览器类型 */
  browser?: string;
  /** 记录ID */
  id?: string;
  /** 操作系统 */
  os?: string;
  /** 分享ID */
  shareId?: string;
}